Move decoding of stacktrace response
