+++ x-pack/plugins/profiling/server/routes/search_stacktraces.test.ts
@@ -5,8 +5,11 @@
  * 2.0.
  */
 
-import { decodeStackTraceResponse, makeFrameID } from './search_stacktraces';
-import { StackTraceResponse } from '../../common/stack_traces';
+import {
+  decodeStackTraceResponse,
+  makeFrameID,
+  StackTraceResponse,
+} from '../../common/stack_traces';
 
 describe('Stack trace response operations', () => {
   test('empty stack trace response', () => {
